requirements.txt
requests>=2.31,<3
tqdm>=4.66,<5
pandas>=2.1,<3
datasets>=2.16,<3
numpy>=1.26,<2
python-dotenv>=1.0,<2
google-genai>=1.0,<2
faiss-cpu>=1.7,<2
